requirements.txt
Flask==2.3.3
google-cloud-storage==2.9.0
gunicorn==21.2.0
PyYAML==6.0.3